Share single Spring context across the schedule integration tests
